feat: now the in-progress todo is linked to a timer.
