+++ src/components/timer/TimerProgressBar.vue
@@ -1,14 +1,16 @@
 <script setup lang="ts">
 import type { TimerProgressBarProps } from '@/types/interfaces/Timer';
+import { useTimer } from '@/composable/useTimer';
 import { useTimerTypeKey } from '@/composable/useTimerTypeKey';
-import { computed } from 'vue';
+import { computed, onUpdated } from 'vue';
 
 const props = defineProps<TimerProgressBarProps>();
 
 const timerCurrentValue = computed(() => props.timerType);
+const { timerSettings } = useTimer();
 const { timerTypeKey } = useTimerTypeKey(timerCurrentValue);
 
-const progressMax = computed(() => props.timerSettings[timerTypeKey.value]);
+const progressMax = computed(() => timerSettings[timerTypeKey.value]);
 const progressValue = computed(() => progressMax.value - props.timer);
 </script>
