+++ src/components/timer/TimerSettingsModal.vue
@@ -3,36 +3,28 @@ import type { TimerSettingsModal } from '@/types/interfaces/Timer';
 import TimerSettingsModalFieldsetSound from '@/components/fieldsets/TimerSettingsModalFieldsetSound.vue';
 import TimerSettingsModalFieldsetTimer from '@/components/fieldsets/TimerSettingsModalFieldsetTimer.vue';
 import BaseInput from '@/components/inputs/BaseInput.vue';
+import { useTimer } from '@/composable/useTimer';
 import { INPUT_LIMITS } from '@/types/enums/InputLimits';
 import { toMinutesFixed } from '@/utils/toMinutesFixed';
 import { computed, onMounted, onUnmounted, reactive, ref, watch } from 'vue';
 
-const props = withDefaults(defineProps<TimerSettingsModal>(), {
-  timerSettings: () =>
-    ({
-      focusDuration: 1800,
-      shortBreakDuration: 300,
-      longBreakDuration: 600,
-      rounds: 3,
-      timerFormat: 'seconds',
-      ringAtTheEnd: false,
-      volume: 0.5,
-    }),
-});
+const props = defineProps<TimerSettingsModal>();
 
 const emit = defineEmits<{
   (e: 'close'): void
   (e: 'update'): void
 }>();
 
+const { timerSettings } = useTimer();
+
 const localTimerSettings = reactive({
-  focusDuration: props.timerSettings.focusDuration,
-  shortBreakDuration: props.timerSettings.shortBreakDuration,
-  longBreakDuration: props.timerSettings.longBreakDuration,
-  rounds: props.timerSettings.rounds,
-  timerFormat: props.timerSettings.timerFormat,
-  ringAtTheEnd: props.timerSettings.ringAtTheEnd,
-  volume: props.timerSettings.volume,
+  focusDuration: timerSettings.focusDuration,
+  shortBreakDuration: timerSettings.shortBreakDuration,
+  longBreakDuration: timerSettings.longBreakDuration,
+  rounds: timerSettings.rounds,
+  timerFormat: timerSettings.timerFormat,
+  ringAtTheEnd: timerSettings.ringAtTheEnd,
+  volume: timerSettings.volume,
 });
 
 const editSettings = () => {
