+++ src/types/interfaces/Timer.ts
@@ -3,7 +3,6 @@ import type { TIMER_STATUS, TIMER_TYPE } from '@/types/enums/Timer';
 export interface TimerProgressBarProps {
   timer: number
   timerType: TIMER_TYPE
-  timerSettings: TimerSettings
 }
 
 export interface TimerSettings {
@@ -17,15 +16,6 @@ export interface TimerSettings {
 }
 
 export interface TimerSettingsModal {
-  timerSettings: {
-    focusDuration: number
-    shortBreakDuration: number
-    longBreakDuration: number
-    rounds: number
-    timerFormat: string
-    ringAtTheEnd: boolean
-    volume: number
-  }
   settingsIconRef: HTMLElement | null
 }
