+++ src/types/interfaces/Todo.ts
@@ -1,6 +1,7 @@
 export interface Todo {
   name: string
   time: number | null
+  timeDone: number
   id: string
   status: string
 }
